Change BigInt class bit-shift operators to accept unsigned long instead of int
This is because libgmp expects unsigned long for the bitcount, not int.
So negative shifts are unsupported by libgmp. Therefore we simply
modified the interface for BigInt class to reflect that it should really
be an unsigned long bitcount. Also modified the newly introduced shift-related
wrapper helpers in class ScriptBigInt to also accept unsigned long.

+++ src/script/bigint.cpp
@@ -566,12 +566,12 @@ BigInt &BigInt::operator^=(const BigInt &o) {
 BigInt &BigInt::operator++() { ++p().base(); return *this; }
 BigInt &BigInt::operator--() { --p().base(); return *this; }
 
-BigInt &BigInt::operator<<=(int x) {
+BigInt &BigInt::operator<<=(unsigned long x) {
     p().base() <<= x;
     return *this;
 }
 
-BigInt &BigInt::operator>>=(int x) {
+BigInt &BigInt::operator>>=(unsigned long x) {
     p().base() >>= x;
     return *this;
 }
